+++ tests/testthat/test-hybrid_generics.R
@@ -1,9 +1,9 @@
 # hybrid methods work ------------------------------------------------------
 
-testthat::test_that(
+test_that(
   desc = "hybrid methods works",
   code = {
-    testthat::skip_on_cran()
+    # skip_on_cran()
 
     library(lme4)
     # merMord
@@ -12,13 +12,13 @@ testthat::test_that(
     df_lmm <- tidy_parameters(lmm_mod, effects = "fixed")
 
     # test
-    testthat::expect_equal(
+    expect_equal(
       df_lmm$estimate,
       c(251.40510, 10.46729),
       tolerance = 0.001
     )
 
-    testthat::expect_equal(
+    expect_equal(
       dim(glance_performance(lmm_mod, effects = "fixed")), c(1L, 10L)
     )
 
@@ -28,9 +28,9 @@ testthat::test_that(
     df_lm <- tidy_parameters(lm_mod, robust = TRUE)
 
     # test
-    testthat::expect_equal(df_lm$estimate[1], 251.4051, tolerance = 0.001)
+    expect_equal(df_lm$estimate[1], 251.4051, tolerance = 0.001)
 
-    testthat::expect_equal(
+    expect_equal(
       dim(glance_performance(lm_mod, effects = "fixed"))[[1]], 1L
     )
 
@@ -54,7 +54,7 @@ testthat::test_that(
       )
 
     # test
-    testthat::expect_identical(
+    expect_identical(
       tidy_parameters(mod_lavaan, effects = "fixed"),
       tidy_parameters(mod_lavaan, effects = "fixed", exponentiate = TRUE)
     )
@@ -66,17 +66,17 @@ testthat::test_that(
     df <- suppressMessages(suppressWarnings(tidy_parameters(tidy(mod))))
 
     # test
-    testthat::expect_equal(dim(df_rlm), c(4L, 8L))
-    testthat::expect_s3_class(df, "tbl_df")
+    expect_equal(dim(df_rlm), c(4L, 9L))
+    expect_s3_class(df, "tbl_df")
 
     # model (not covered by easystats but covered by broom)
-    testthat::expect_equal(
+    expect_equal(
       dim(suppressWarnings(broomExtra::tidy_parameters(acf(lh, plot = FALSE)))),
       c(17L, 2L)
     )
 
     # where functions are not supposed to work
-    testthat::expect_null(suppressWarnings(broomExtra::glance_performance(acf(lh, plot = FALSE))))
-    testthat::expect_null(broomExtra::tidy_parameters(list(1, c("x", "y")), verbose = FALSE))
+    expect_null(suppressWarnings(broomExtra::glance_performance(acf(lh, plot = FALSE))))
+    expect_null(broomExtra::tidy_parameters(list(1, c("x", "y")), verbose = FALSE))
   }
 )